Change to CTC timer mode and implement one motor
Timer is in CTC timer mode which means that when the value of OCR1A is
reached, it triggers the interruption and reset the timer to 0. That's
easier to use because you always know how long the timer last.

One motor has been implemented (simulated by a led). It now need a queue
of command for the different motor to move.